+++ js/yjs.js
@@ -136,95 +136,6 @@ export const createSystem = async () => {
   return system;
 };
 
-// Get character data from Yjs
-export const getCharacter = () => {
-  if (!yjsSystem?.journalMap) return { name: '', race: '', class: '', backstory: '', notes: '' };
-  
-  const characterMap = yjsSystem.journalMap.get('character');
-  if (!characterMap) {
-    return { name: '', race: '', class: '', backstory: '', notes: '' };
-  }
-  
-  return {
-    name: characterMap.get('name') || '',
-    race: characterMap.get('race') || '',
-    class: characterMap.get('class') || '',
-    backstory: characterMap.get('backstory') || '',
-    notes: characterMap.get('notes') || ''
-  };
-};
-
-// Update character data
-export const updateCharacter = (characterData) => {
-  if (!yjsSystem?.journalMap) return;
-  
-  let characterMap = yjsSystem.journalMap.get('character');
-  if (!characterMap) {
-    characterMap = new Y.Map();
-    yjsSystem.journalMap.set('character', characterMap);
-  }
-  
-  // Set each field individually for CRDT conflict resolution
-  characterMap.set('name', characterData.name || '');
-  characterMap.set('race', characterData.race || '');
-  characterMap.set('class', characterData.class || '');
-  characterMap.set('backstory', characterData.backstory || '');
-  characterMap.set('notes', characterData.notes || '');
-  
-  yjsSystem.journalMap.set('lastModified', Date.now());
-};
-
-// Get settings from Yjs
-export const getSettings = () => {
-  if (!yjsSystem?.settingsMap) return { apiKey: '', enableAIFeatures: false };
-  
-  return {
-    apiKey: yjsSystem.settingsMap.get('apiKey') || '',
-    enableAIFeatures: yjsSystem.settingsMap.get('enableAIFeatures') || false
-  };
-};
-
-// Update settings
-export const updateSettings = (settings) => {
-  if (!yjsSystem?.settingsMap) return;
-  
-  yjsSystem.settingsMap.set('apiKey', settings.apiKey || '');
-  yjsSystem.settingsMap.set('enableAIFeatures', Boolean(settings.enableAIFeatures));
-};
-
-// Get summaries from Yjs
-export const getSummaries = () => {
-  if (!yjsSystem?.summariesMap) return {};
-  
-  const summaries = {};
-  yjsSystem.summariesMap.forEach((summaryMap, key) => {
-    summaries[key] = {
-      content: summaryMap.get('content') || '',
-      words: summaryMap.get('words') || 0,
-      timestamp: summaryMap.get('timestamp') || Date.now()
-    };
-  });
-  
-  return summaries;
-};
-
-// Update summaries
-export const updateSummaries = (summaries) => {
-  if (!yjsSystem?.summariesMap) return;
-  
-  // Clear existing summaries
-  yjsSystem.summariesMap.clear();
-  
-  // Add each summary
-  Object.entries(summaries).forEach(([key, summary]) => {
-    const summaryMap = new Y.Map();
-    summaryMap.set('content', summary.content || summary.summary || '');
-    summaryMap.set('words', summary.words || 0);
-    summaryMap.set('timestamp', summary.timestamp || Date.now());
-    yjsSystem.summariesMap.set(key, summaryMap);
-  });
-};
-
 // Get sync status
 export const getSyncStatus = (providers) => {
   if (!providers || providers.length === 0) {
